+++ modules/codec/ffmpeg/encoder.c
@@ -757,9 +757,9 @@ static block_t *EncodeVideo( encoder_t *p_enc, picture_t *p_pict )
     if( 1 )
 #endif
     {
-        frame.pts = p_pict->date ? p_pict->date : AV_NOPTS_VALUE;
+        frame.pts = p_pict->date ? p_pict->date : (signed int) AV_NOPTS_VALUE;
 
-        if ( p_sys->b_hurry_up && frame.pts != AV_NOPTS_VALUE )
+        if ( p_sys->b_hurry_up && frame.pts != (signed int) AV_NOPTS_VALUE )
         {
             mtime_t current_date = mdate();
 
@@ -805,7 +805,7 @@ static block_t *EncodeVideo( encoder_t *p_enc, picture_t *p_pict )
         frame.pts = AV_NOPTS_VALUE;
     }
 
-    if ( frame.pts != AV_NOPTS_VALUE && frame.pts != 0 )
+    if ( frame.pts != (signed int) AV_NOPTS_VALUE && frame.pts != 0 )
     {
         if ( p_sys->i_last_pts == frame.pts )
         {
@@ -857,7 +857,7 @@ static block_t *EncodeVideo( encoder_t *p_enc, picture_t *p_pict )
             /* No delay -> output pts == input pts */
             p_block->i_pts = p_block->i_dts = p_pict->date;
         }
-        else if( p_sys->p_context->coded_frame->pts != AV_NOPTS_VALUE &&
+        else if( p_sys->p_context->coded_frame->pts != (signed int) AV_NOPTS_VALUE &&
             p_sys->p_context->coded_frame->pts != 0 &&
             p_sys->i_buggy_pts_detect != p_sys->p_context->coded_frame->pts )
         {
